Completely reworked the synchronization code
Recovery from a desychronization is not possible with fewer than two memory-
mapped buffer. When attempting to recover from desynchronization, the following
events occur (assuming a LAGL):

	1. Notice a LAGL condition (i.e. left + err < right).
    2. Drop left frame.
	3. Block while waiting for a new left frame.
	4. Copy the frame to user memory when available:
	5.a. Dequeue the single kernel buffer.
	5.b. Next left frame is dropped (no buffers available)!
	5.c. Copy the frame to user memory.
	5.d. Enqueue the buffer for reuse.
	6. Notice a LAGR condition (i.e. left - err > right).
	7. Repeat from step 1.

This can be solved by simply increasing the number of kernel buffers for each
camera to a minimum of two. Because the buffering is handled in the kernel,
the CameraMonitor class is completely superfluous and has been removed.