Fixed bug in Webcam::SetResolution(); now working
Previously decreasing the image resolution would work as expected, but
increasing resolution would fail. This occured because the kernel buffers for
memory-mapped IO were allocated in the constructor using the camera's default
resolution. Increasing the resolution from the default would require larger
buffers than those allocated and causes the Video4Linux driver to reject the
negotiation.

This was fixed by moving the buffer allocation code out of the constructor
and into Webcam::SetStreaming(). Currently, this code reallocates all of the
kernel buffers each time streaming is started (regardless of whether the
resolution has changed). This is obviously non-ideal and should be fixed in
a later commit.